+++ DatadogRUM/Sources/RUMMonitor/Scopes/RUMViewScope.swift
@@ -68,7 +68,13 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
     /// Tells if this View is the active one.
     /// `true` for every new started View.
     /// `false` if the View was stopped or any other View was started.
-    private(set) var isActiveView = true
+    private(set) var isActiveView = true {
+        didSet {
+            if oldValue && !isActiveView {
+                networkSettledMetric.trackViewWasStopped()
+            }
+        }
+    }
     /// Tells if this scope has received the "start" command.
     /// If `didReceiveStartCommand == true` and another "start" command is received for this View this scope is marked as inactive.
     private var didReceiveStartCommand = false
@@ -97,6 +103,11 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
 
     private var viewPerformanceMetrics: [PerformanceMetric: VitalInfo] = [:]
 
+    /// Time-to-Network-Settled metric for this view.
+    private let networkSettledMetric: TTNSMetricTracking
+    /// Interaction-to-Next-View metric for this view.
+    private let interactionToNextViewMetric: ITNVMetricTracking
+
     init(
         isInitialView: Bool,
         parent: RUMContextProvider,
@@ -106,7 +117,8 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
         name: String,
         customTimings: [String: Int64],
         startTime: Date,
-        serverTimeOffset: TimeInterval
+        serverTimeOffset: TimeInterval,
+        interactionToNextViewMetric: ITNVMetricTracking
     ) {
         self.parent = parent
         self.dependencies = dependencies
@@ -118,6 +130,7 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
         self.viewName = name
         self.viewStartTime = startTime
         self.serverTimeOffset = serverTimeOffset
+        self.interactionToNextViewMetric = interactionToNextViewMetric
 
         self.vitalInfoSampler = dependencies.vitalsReaders.map {
             .init(
@@ -127,6 +140,8 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
                 frequency: $0.frequency
             )
         }
+        self.networkSettledMetric = dependencies.networkSettledMetricFactory(viewStartTime, viewName)
+        interactionToNextViewMetric.trackViewStart(at: startTime, name: name, viewID: viewUUID)
 
         // Notify Synthetics if needed
         if dependencies.syntheticsTest != nil && self.context.sessionID != .nullUUID {
@@ -264,6 +279,10 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
         let hasNoPendingResources = resourceScopes.isEmpty
         let shouldComplete = !isActiveView && hasNoPendingResources
 
+        if shouldComplete {
+            interactionToNextViewMetric.trackViewComplete(viewID: viewUUID)
+        }
+
         return !shouldComplete
     }
 
@@ -296,12 +315,17 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
             httpMethod: command.httpMethod,
             resourceKindBasedOnRequest: command.kind,
             spanContext: command.spanContext,
-            onResourceEventSent: { [weak self] in
-                self?.resourcesCount += 1
+            networkSettledMetric: networkSettledMetric,
+            onResourceEvent: { [weak self] wasSent in
+                if wasSent {
+                    self?.resourcesCount += 1
+                }
                 self?.needsViewUpdate = true
             },
-            onErrorEventSent: { [weak self] in
-                self?.errorsCount += 1
+            onErrorEvent: { [weak self] wasSent in
+                if wasSent {
+                    self?.errorsCount += 1
+                }
                 self?.needsViewUpdate = true
             }
         )
@@ -318,6 +342,7 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
             serverTimeOffset: serverTimeOffset,
             isContinuous: true,
             instrumentation: command.instrumentation,
+            interactionToNextViewMetric: interactionToNextViewMetric,
             onActionEventSent: { [weak self] event in
                 self?.onActionEventSent(event)
             }
@@ -335,6 +360,7 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
             serverTimeOffset: serverTimeOffset,
             isContinuous: false,
             instrumentation: command.instrumentation,
+            interactionToNextViewMetric: interactionToNextViewMetric,
             onActionEventSent: { [weak self] event in
                 self?.onActionEventSent(event)
             }
@@ -490,6 +516,8 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
         let memoryInfo = vitalInfoSampler?.memory
         let refreshRateInfo = vitalInfoSampler?.refreshRate
         let isSlowRendered = refreshRateInfo?.meanValue.map { $0 < Constants.slowRenderingThresholdFPS }
+        let networkSettledTime = networkSettledMetric.value(at: command.time, appStateHistory: context.applicationStateHistory)
+        let interactionToNextViewTime = interactionToNextViewMetric.value(for: viewUUID)
 
         let viewEvent = RUMViewEvent(
             dd: .init(
@@ -565,7 +593,7 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
                 interactionToNextPaint: nil,
                 interactionToNextPaintTargetSelector: nil,
                 interactionToNextPaintTime: nil,
-                interactionToNextViewTime: nil,
+                interactionToNextViewTime: interactionToNextViewTime?.toInt64Nanoseconds,
                 isActive: isActive,
                 isSlowRendered: isSlowRendered ?? false,
                 jsRefreshRate: viewPerformanceMetrics[.jsFrameTimeSeconds]?.asJsRefreshRate(),
@@ -578,7 +606,7 @@ internal class RUMViewScope: RUMScope, RUMContextProvider {
                 memoryAverage: memoryInfo?.meanValue,
                 memoryMax: memoryInfo?.maxValue,
                 name: viewName,
-                networkSettledTime: nil,
+                networkSettledTime: networkSettledTime?.toInt64Nanoseconds,
                 referrer: nil,
                 refreshRateAverage: refreshRateInfo?.meanValue,
                 refreshRateMin: refreshRateInfo?.minValue,