+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 - [FEATURE] Adds anonymous identifier configuration for RUM Sessions linking. See [#2172][]
 
+- [FEATURE] Add Time To Network Setled metric in RUM. See [#2125][]
+- [FEATURE] Add Interaction To Next View metric in RUM. See [#2153][]
 - [FIX] Fix SwiftUI placeholder in Session Replay when Feature Flag is disabled. See [#2170][]
 
 # 2.22.0 / 02-01-2025
@@ -810,10 +812,12 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2063]: https://github.com/DataDog/dd-sdk-ios/pull/2063
 [#2092]: https://github.com/DataDog/dd-sdk-ios/pull/2092
 [#2113]: https://github.com/DataDog/dd-sdk-ios/pull/2113
+[#2125]: https://github.com/DataDog/dd-sdk-ios/pull/2125
 [#2114]: https://github.com/DataDog/dd-sdk-ios/pull/2114
 [#2116]: https://github.com/DataDog/dd-sdk-ios/pull/2116
 [#2120]: https://github.com/DataDog/dd-sdk-ios/pull/2120
 [#2126]: https://github.com/DataDog/dd-sdk-ios/pull/2126
+[#2153]: https://github.com/DataDog/dd-sdk-ios/pull/2153
 [#2148]: https://github.com/DataDog/dd-sdk-ios/pull/2148
 [#2154]: https://github.com/DataDog/dd-sdk-ios/pull/2154
 [#2172]: https://github.com/DataDog/dd-sdk-ios/pull/2172